split stats into seperate scripts
ready to introduce some mechanics for rate of decay, conservation conditions and so on

+++ Assets/Scripts/Environment/IHeatable.cs
@@ -2,6 +2,6 @@
 {
     public interface IHeatable
     {
-        public void Heat(int heatStrength);
+        public void Heat(float heatStrength);
     }
 }
