+++ Assets/Scripts/UI/StatsPanel/PlayerStatsComponent.cs
@@ -11,9 +11,9 @@ namespace TPP.Scripts.UI.Components
         public Image statIcon;
         public Slider statSlider;
 
-        public int maxValue;
+        public float maxValue;
 
-        public void Init(string labelText, int value)
+        public void Init(string labelText, float value)
         {
             statLabel.text = labelText;
             maxValue = value;
@@ -21,7 +21,7 @@ namespace TPP.Scripts.UI.Components
             statSlider.value = CommonUtils.Map(value, 0, maxValue, 0f, 1f);
         }
 
-        public void UpdateValue(int value)
+        public void UpdateValue(float value)
         {
             statSlider.value = CommonUtils.Map(value, 0, maxValue, 0f, 1f);
         }
